Re-org docker list code
