+++ packages/dappmanager/src/calls/autoUpdateDataGet.ts
@@ -1,5 +1,5 @@
 import semver from "semver";
-import { listPackages } from "../modules/docker/listContainers";
+import { listPackages } from "../modules/docker/list";
 import { getCoreVersionId } from "../utils/coreVersionId";
 import * as autoUpdateHelper from "../utils/autoUpdateHelper";
 import { shortNameCapitalized } from "../utils/format";
