+++ packages/dappmanager/src/calls/fetchDnpRequest.ts
@@ -1,6 +1,6 @@
 import { mapValues, omit } from "lodash";
 import semver from "semver";
-import { listPackages } from "../modules/docker/listContainers";
+import { listPackages } from "../modules/docker/list";
 import params from "../params";
 import shouldUpdate from "../modules/dappGet/utils/shouldUpdate";
 import deepmerge from "deepmerge";
