+++ packages/dappmanager/src/calls/packageRestart.ts
@@ -1,5 +1,5 @@
 import * as eventBus from "../eventBus";
-import { listPackage } from "../modules/docker/listContainers";
+import { listPackage } from "../modules/docker/list";
 import { restartPackage } from "../modules/docker/restartPackage";
 import { getDockerTimeoutMax } from "../modules/docker/utils";
