+++ packages/dappmanager/src/calls/packageSetPortMappings.ts
@@ -1,6 +1,6 @@
 import * as eventBus from "../eventBus";
 import params from "../params";
-import { listPackage } from "../modules/docker/listContainers";
+import { listPackage } from "../modules/docker/list";
 import { ComposeFileEditor } from "../modules/compose/editor";
 import { PortMapping } from "../types";
 import { restartPackage } from "../modules/docker/restartPackage";
