+++ packages/dappmanager/src/modules/docker/volumesData.ts
@@ -1,5 +1,5 @@
 import { dockerDf, dockerVolumesList } from "./dockerApi";
-import { listContainers } from "./listContainers";
+import { listContainers } from "./list";
 import { parseDevicePath } from "../compose";
 import { VolumeData, VolumeOwnershipData, PackageContainer } from "../../types";
 import { detectMountpoints } from "../hostScripts";
