+++ packages/dappmanager/src/modules/ethClient/clientStatus.ts
@@ -2,7 +2,7 @@ import { ethers } from "ethers";
 import * as db from "../../db";
 import { ethClientData } from "../../params";
 import { EthClientStatus, EthClientTargetPackage } from "../../types";
-import { listPackageNoThrow } from "../../modules/docker/listContainers";
+import { listPackageNoThrow } from "../../modules/docker/list";
 import { serializeError } from "./types";
 import { getEthClientApiUrl } from "./apiUrl";
 import { parseEthersSyncing } from "../../utils/ethers";
